# server/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "基于 Gin + Vue3 的后台管理系统 API。\u003cbr/\u003e统一响应格式：\u003ccode\u003e{\"code\":0,\"msg\":\"ok\",\"data\":...}\u003c/code\u003e，code=0 成功，code=7 业务失败。\u003cbr/\u003e所有 \u003ccode\u003e/api/v1/system/*\u003c/code\u003e 接口需要 \u003ccode\u003eAuthorization: Bearer \u0026lt;token\u0026gt;\u003c/code\u003e。",
        "title": "go-admin API",
        "contact": {},
        "version": "1.0.0"
    },
    "host": "localhost:8989",
    "basePath": "/",
    "paths": {
...
        "/api/v1/base/info": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "基础接口"
                ],
                "summary": "获取当前登录用户信息",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/github_com_lihongsheng_go-admin_server_utils_response.Body"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/github_com_lihongsheng_go-admin_server_model_system.SysUser"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/github_com_lihongsheng_go-admin_server_utils_response.Body"
                        }
                    }
                }
            }
        },
        "/api/v1/base/login": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "基础接口"
                ],
                "summary": "用户登录",
                "parameters": [
                    {
                        "description": "登录参数",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_lihongsheng_go-admin_server_dto_base.LoginReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/github_com_lihongsheng_go-admin_server_utils_response.Body"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/github_com_lihongsheng_go-admin_server_dto_base.LoginResp"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/api/v1/base/logout": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "基础接口"
                ],
                "summary": "退出登录",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_lihongsheng_go-admin_server_utils_response.Body"
                        }
                    }
                }
            }
        },
        "/api/v1/base/menu": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "基础接口"
                ],
                "summary": "获取当前用户菜单树",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/github_com_lihongsheng_go-admin_server_utils_response.Body"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "object",
                                            "properties": {
                                                "menus": {
                                                    "type": "array"
                                                }
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/github_com_lihongsheng_go-admin_server_utils_response.Body"
                        }
                    }
                }
            }
        },
...
        "/api/v1/system/mch": {
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "商户管理"
                ],
                "summary": "更新商户",
                "parameters": [
                    {
                        "description": "商户信息（需包含 ID）",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_lihongsheng_go-admin_server_dto_system.MchCreateRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_lihongsheng_go-admin_server_utils_response.Body"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "商户管理"
                ],
                "summary": "新增商户",
                "parameters": [
                    {
                        "description": "商户信息",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_lihongsheng_go-admin_server_dto_system.MchCreateRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_lihongsheng_go-admin_server_utils_response.Body"
                        }
                    }
                }
            }
        },
        "/api/v1/system/mch/list": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "商户管理"
                ],
                "summary": "商户列表",
                "parameters": [
                    {
                        "type": "string",
                        "description": "公司名称",
                        "name": "mch_name",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "商户编号",
                        "name": "mch_no",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "状态（1正常 2停用）",
                        "name": "status",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "每页条数（最大50）",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "页码",
                        "name": "page",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/github_com_lihongsheng_go-admin_server_utils_response.Body"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "object",
                                            "properties": {
                                                "list": {
                                                    "type": "array"
                                                },
                                                "total": {
                                                    "type": "integer"
                                                }
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/api/v1/system/mch/no/{mchNo}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "商户管理"
                ],
                "summary": "根据编号获取商户详情",
                "parameters": [
                    {
                        "type": "string",
                        "description": "商户编号",
                        "name": "mchNo",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/github_com_lihongsheng_go-admin_server_utils_response.Body"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/github_com_lihongsheng_go-admin_server_model_system.Merchant"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/api/v1/system/mch/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "商户管理"
                ],
                "summary": "根据 ID 获取商户详情",
                "parameters": [
                    {
                        "type": "integer",
                        "format": "int64",
                        "description": "商户ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/github_com_lihongsheng_go-admin_server_utils_response.Body"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/github_com_lihongsheng_go-admin_server_model_system.Merchant"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "enum.MchStatus": {
            "type": "integer",
            "format": "int32",
            "enum": [
                2,
                1
            ],
            "x-enum-comments": {
                "MchStatusDisabled": "禁用",
                "MchStatusEnabled": "启用"
            },
            "x-enum-descriptions": [
                "禁用",
                "启用"
            ],
            "x-enum-varnames": [
                "MchStatusDisabled",
                "MchStatusEnabled"
            ]
        },
...
        "github_com_lihongsheng_go-admin_server_dto_base.LoginReq": {
            "type": "object",
            "required": [
                "captcha_code",
                "captcha_id",
                "password",
                "username"
            ],
            "properties": {
                "captcha_code": {
                    "type": "string"
                },
                "captcha_id": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        },
        "github_com_lihongsheng_go-admin_server_dto_base.LoginResp": {
            "type": "object",
            "properties": {
                "token": {
                    "type": "string"
                },
                "user": {
                    "$ref": "#/definitions/github_com_lihongsheng_go-admin_server_model_system.SysUser"
                }
            }
        },
        "github_com_lihongsheng_go-admin_server_dto_system.MchCreateRequest": {
            "type": "object",
            "properties": {
                "address": {
                    "description": "地址",
                    "type": "string"
                },
                "email": {
                    "description": "联系电话",
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "linker": {
                    "description": "联系人",
                    "type": "string"
                },
                "mch_name": {
                    "description": "公司名字",
                    "type": "string"
                },
                "phone": {
                    "description": "联系电话",
                    "type": "string"
                },
                "reason": {
                    "description": "备注",
                    "type": "string"
                },
                "status": {
                    "description": "0 停用 1 正常",
                    "allOf": [
                        {
                            "$ref": "#/definitions/enum.MchStatus"
                        }
                    ]
                }
            }
        },
...
        "github_com_lihongsheng_go-admin_server_model_system.Merchant": {
            "type": "object",
            "properties": {
                "address": {
                    "description": "地址",
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "email": {
                    "description": "联系电话",
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "linker": {
                    "description": "联系人",
                    "type": "string"
                },
                "mch_name": {
                    "description": "公司名字",
                    "type": "string"
                },
                "mch_no": {
                    "description": "编号",
                    "type": "string"
                },
                "phone": {
                    "description": "联系电话",
                    "type": "string"
                },
                "reason": {
                    "description": "备注",
                    "type": "string"
                },
                "status": {
                    "description": "2 停用 1 正常",
                    "type": "integer"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "github_com_lihongsheng_go-admin_server_model_system.SysMenu": {
            "type": "object",
            "properties": {
                "api_rules": {
                    "description": "JSON 数组：[{\"path\":\"/api/...\",\"method\":\"GET\"},...]",
                    "type": "string"
                },
                "children": {
                    "description": "虚拟字段，仅用于 JSON 序列化时组装树",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_lihongsheng_go-admin_server_model_system.SysMenu"
                    }
                },
                "component": {
                    "description": "前端组件路径；catalog 用 \"Layout\"",
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "hidden": {
                    "description": "是否隐藏",
                    "type": "boolean"
                },
                "icon": {
                    "description": "图标名（element-ui icon）",
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "keep_alive": {
                    "description": "是否缓存页面",
                    "type": "boolean"
                },
                "name": {
                    "description": "路由 name；button 为按钮名",
                    "type": "string"
                },
                "parent_id": {
                    "description": "父节点 ID，0 表示根",
                    "type": "integer"
                },
                "path": {
                    "description": "路由路径；catalog 如 /system，menu 如 user",
                    "type": "string"
                },
                "permission": {
                    "description": "权限码；button 必有值，如 user:add",
                    "type": "string"
                },
                "redirect": {
                    "description": "重定向路径",
                    "type": "string"
                },
                "sort": {
                    "description": "排序序号",
                    "type": "integer"
                },
                "title": {
                    "description": "显示标题",
                    "type": "string"
                },
                "type": {
                    "description": "catalog / menu / button",
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "github_com_lihongsheng_go-admin_server_model_system.SysRole": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "default_router": {
                    "description": "登录后默认首页路由",
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "menus": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_lihongsheng_go-admin_server_model_system.SysMenu"
                    }
                },
                "name": {
                    "type": "string"
                },
                "remark": {
                    "type": "string"
                },
                "status": {
                    "type": "integer"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "github_com_lihongsheng_go-admin_server_model_system.SysUser": {
            "type": "object",
            "properties": {
                "avatar": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "nickname": {
                    "type": "string"
                },
                "phone": {
                    "type": "string"
                },
                "roles": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_lihongsheng_go-admin_server_model_system.SysRole"
                    }
                },
                "status": {
                    "description": "1启用 0禁用",
                    "type": "integer"
                },
                "updated_at": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        },
        "github_com_lihongsheng_go-admin_server_utils_response.Body": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {},
                "msg": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}